backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
python-multipart==0.0.6
pypdf2==3.0.1
python-docx==1.1.0
chromadb==0.4.22
openai==1.10.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.5.3
pydantic-settings==2.1.0
aiosqlite==0.19.0
